server/requirements.txt
asgiref==3.7.2
coverage==7.2.7
Django==4.2.2
django-axes==6.0.5
django-cors-headers==4.1.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.11.2
mysqlclient==2.1.1
Pillow==9.5.0
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
